+++ lib/core/providers/game_state_providers.dart
@@ -3,6 +3,8 @@ import 'dart:async';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:uuid/uuid.dart';
 import '../models/game_state.dart';
+import '../models/player_profile.dart';
+import '../utils/title_evaluator.dart';
 import 'player_providers.dart';
 import 'repository_providers.dart';
 import 'package:dgv/features/firebase/providers/firebase_providers.dart';
@@ -92,7 +94,7 @@ class GameStateNotifier extends _$GameStateNotifier {
     final syncService = ref.read(firebaseSyncServiceProvider);
     unawaited(
       playerRepo.getAll().then(
-        (p) => syncService.syncGameFinish(updatedState, p),
+        (p) => syncService.syncGameFinish(updatedState, p, _buildCeremony(p)),
       ),
     );
     // Flush player cache so post-game screens (main menu, ceremony) see
@@ -114,4 +116,51 @@ class GameStateNotifier extends _$GameStateNotifier {
     await repository.save(state);
     ref.invalidateSelf();
   }
+}
+
+const _kStickerKeys = [
+  'sin_pegatina',
+  'pegatina_b',
+  'pegatina_c',
+  'pegatina_eco',
+  'pegatina_0_emisiones',
+];
+
+Map<String, dynamic> _buildCeremony(List<PlayerProfile> players) {
+  final podium = TitleEvaluator.calculateLeaderboard(players).take(3).toList();
+  final coleccionista = TitleEvaluator.getMostTitlesPlayer(players);
+  final environmentals = TitleEvaluator.getEnvironmentalDistinctives(players);
+
+  return {
+    'podium': [
+      for (var i = 0; i < podium.length; i++)
+        {
+          'rank': i + 1,
+          'playerId': podium[i].id,
+          'name': podium[i].name,
+          'surname': podium[i].surname,
+          'points': podium[i].points,
+        },
+    ],
+    'coleccionista': coleccionista == null
+        ? null
+        : {
+            'playerId': coleccionista.id,
+            'name': coleccionista.name,
+            'surname': coleccionista.surname,
+            'totalTitles': coleccionista.titleCounts.values
+                .fold(0, (s, c) => s + c),
+          },
+    'environmentals': [
+      for (var i = 0; i < environmentals.length; i++)
+        {
+          'rank': i + 1,
+          'stickerKey': _kStickerKeys[i],
+          'playerId': environmentals[i].id,
+          'name': environmentals[i].name,
+          'surname': environmentals[i].surname,
+          'maxBAC': environmentals[i].maxBAC,
+        },
+    ],
+  };
 }